feat: (#488) 인기글 랭킹 목록 데이터 프롭스로 넘겨주는 방식으로 수정
