+++ frontend/src/pages/Ranking/PopularPost/index.tsx
@@ -19,20 +19,12 @@ const rankIconUrl: Record<number, string> = {
 };
 
 const columnNameList = ['등수', '닉네임', '글 제목', '투표 수'];
-const rankingPostInfo: RankingPost = {
-  rank: 1,
-  post: {
-    id: 29,
-    nickname: 'writer',
-    title: '이것은 엄청나게 많은 투표가 이루어진 대단한 글이지요',
-    voteCount: 10,
-  },
-};
-const rankingPostList: RankingPost[] = new Array(10)
-  .fill(rankingPostInfo)
-  .map((post, index) => ({ ...post, ranking: index + 1 }));
 
-export default function PopularPost() {
+interface PopularPostProps {
+  rankingPostList: RankingPost[];
+}
+
+export default function PopularPost({ rankingPostList }: PopularPostProps) {
   return (
     <RS.Background>
       <S.Table>
